Refactors posts-component structure.
