+++ components/posts-list.tsx
@@ -1,20 +1,28 @@
-import { Post } from "@/lib/types";
+import { PostListResult } from "@/lib/types";
 import PostItem, { CardType, PostItemLoading, ProfilePostItem, ProfilePostItemLoading } from "./post-item";
+import { PaginationPaging } from "./pagination";
 
-export default function PostsList({ posts, card }: { posts: Post[], card?: CardType }) {
+export default async function PostsList({ data, card }: { data: Promise<PostListResult>, card?: CardType }) {
+
+    const result = await data;
+    const paging = result.data?.meta;
+    const posts = result.data?.posts ?? [];
 
     return (
-        <ul className="grid grid-cols-1 auto-rows-min gap-4">
-            {posts.map(post => (
-                <li key={post.id}>
-                    {(card && card === CardType.Profile) ?
-                        <ProfilePostItem post={post} />
-                        :
-                        <PostItem post={post} />
-                    }
-                </li>
-            ))}
-        </ul>
+        <>
+            <ul className="grid grid-cols-1 auto-rows-min gap-4">
+                {posts.map(post => (
+                    <li key={post.id}>
+                        {(card && card === CardType.Profile) ?
+                            <ProfilePostItem post={post} />
+                            :
+                            <PostItem post={post} />
+                        }
+                    </li>
+                ))}
+            </ul>
+            <PaginationPaging params={paging} />
+        </>
     );
 }
