+++ components/profile-content.tsx
@@ -1,12 +1,9 @@
 import { PostListParams, Profile } from "@/lib/types";
 import { Avatar, AvatarFallback, AvatarImage } from "./ui/avatar";
 import { ProfileTag } from "./tags";
-import { PostsService } from "@/lib/service/posts";
-import { Suspense } from "react";
 import PostsFeed from "./posts-feed";
 import { CardType } from "./post-item";
 import { ProfilePostsFilter } from "./posts-filter";
-import { PostsListLoading } from "./posts-list";
 
 export default async function ProfileContent({ data, postParams }: { data: Promise<Profile | null>, postParams: PostListParams }) {
     const profile = await data;
@@ -18,7 +15,6 @@ export default async function ProfileContent({ data, postParams }: { data: Promi
     if (postParams) {
         postParams.profileId = profile.id;
     }
-    const postData = PostsService().getPosts(postParams);
 
     return (
         <>
@@ -65,9 +61,7 @@ export default async function ProfileContent({ data, postParams }: { data: Promi
                     <h2 className="text-xl">Posts</h2>
                     <ProfilePostsFilter />
                 </div>
-                <Suspense fallback={<PostsListLoading card={CardType.Profile} />}>
-                    <PostsFeed data={postData} card={CardType.Profile} />
-                </Suspense>
+                <PostsFeed params={postParams} card={CardType.Profile} />
             </section>
         </>
     );
